- Added new disclaimer gui - Added superclass for hyperium-based GUI's (woohoo)
