+++ src/main/java/cc/hyperium/gui/GuiHyperiumScreenMainMenu.java
@@ -1,0 +1,463 @@
+/*
+ *                    GNU LESSER GENERAL PUBLIC LICENSE
+ *                        Version 3, 29 June 2007
+ *
+ *  Copyright (C) 2007 Free Software Foundation, Inc. <http://fsf.org/>
+ *  Everyone is permitted to copy and distribute verbatim copies
+ *  of this license document, but changing it is not allowed.
+ *
+ *
+ *   This version of the GNU Lesser General Public License incorporates
+ * the terms and conditions of version 3 of the GNU General Public
+ * License, supplemented by the additional permissions listed below.
+ *
+ *   0. Additional Definitions.
+ *
+ *   As used herein, "this License" refers to version 3 of the GNU Lesser
+ * General Public License, and the "GNU GPL" refers to version 3 of the GNU
+ * General Public License.
+ *
+ *   "The Library" refers to a covered work governed by this License,
+ * other than an Application or a Combined Work as defined below.
+ *
+ *   An "Application" is any work that makes use of an interface provided
+ * by the Library, but which is not otherwise based on the Library.
+ * Defining a subclass of a class defined by the Library is deemed a mode
+ * of using an interface provided by the Library.
+ *
+ *   A "Combined Work" is a work produced by combining or linking an
+ * Application with the Library.  The particular version of the Library
+ * with which the Combined Work was made is also called the "Linked
+ * Version".
+ *
+ *   The "Minimal Corresponding Source" for a Combined Work means the
+ * Corresponding Source for the Combined Work, excluding any source code
+ * for portions of the Combined Work that, considered in isolation, are
+ * based on the Application, and not on the Linked Version.
+ *
+ *   The "Corresponding Application Code" for a Combined Work means the
+ * object code and/or source code for the Application, including any data
+ * and utility programs needed for reproducing the Combined Work from the
+ * Application, but excluding the System Libraries of the Combined Work.
+ *
+ *   1. Exception to Section 3 of the GNU GPL.
+ *
+ *   You may convey a covered work under sections 3 and 4 of this License
+ * without being bound by section 3 of the GNU GPL.
+ *
+ *   2. Conveying Modified Versions.
+ *
+ *   If you modify a copy of the Library, and, in your modifications, a
+ * facility refers to a function or data to be supplied by an Application
+ * that uses the facility (other than as an argument passed when the
+ * facility is invoked), then you may convey a copy of the modified
+ * version:
+ *
+ *    a) under this License, provided that you make a good faith effort to
+ *    ensure that, in the event an Application does not supply the
+ *    function or data, the facility still operates, and performs
+ *    whatever part of its purpose remains meaningful, or
+ *
+ *    b) under the GNU GPL, with none of the additional permissions of
+ *    this License applicable to that copy.
+ *
+ *   3. Object Code Incorporating Material from Library Header Files.
+ *
+ *   The object code form of an Application may incorporate material from
+ * a header file that is part of the Library.  You may convey such object
+ * code under terms of your choice, provided that, if the incorporated
+ * material is not limited to numerical parameters, data structure
+ * layouts and accessors, or small macros, inline functions and templates
+ * (ten or fewer lines in length), you do both of the following:
+ *
+ *    a) Give prominent notice with each copy of the object code that the
+ *    Library is used in it and that the Library and its use are
+ *    covered by this License.
+ *
+ *    b) Accompany the object code with a copy of the GNU GPL and this license
+ *    document.
+ *
+ *   4. Combined Works.
+ *
+ *   You may convey a Combined Work under terms of your choice that,
+ * taken together, effectively do not restrict modification of the
+ * portions of the Library contained in the Combined Work and reverse
+ * engineering for debugging such modifications, if you also do each of
+ * the following:
+ *
+ *    a) Give prominent notice with each copy of the Combined Work that
+ *    the Library is used in it and that the Library and its use are
+ *    covered by this License.
+ *
+ *    b) Accompany the Combined Work with a copy of the GNU GPL and this license
+ *    document.
+ *
+ *    c) For a Combined Work that displays copyright notices during
+ *    execution, include the copyright notice for the Library among
+ *    these notices, as well as a reference directing the user to the
+ *    copies of the GNU GPL and this license document.
+ *
+ *    d) Do one of the following:
+ *
+ *        0) Convey the Minimal Corresponding Source under the terms of this
+ *        License, and the Corresponding Application Code in a form
+ *        suitable for, and under terms that permit, the user to
+ *        recombine or relink the Application with a modified version of
+ *        the Linked Version to produce a modified Combined Work, in the
+ *        manner specified by section 6 of the GNU GPL for conveying
+ *        Corresponding Source.
+ *
+ *        1) Use a suitable shared library mechanism for linking with the
+ *        Library.  A suitable mechanism is one that (a) uses at run time
+ *        a copy of the Library already present on the user's computer
+ *        system, and (b) will operate properly with a modified version
+ *        of the Library that is interface-compatible with the Linked
+ *        Version.
+ *
+ *    e) Provide Installation Information, but only if you would otherwise
+ *    be required to provide such information under section 6 of the
+ *    GNU GPL, and only to the extent that such information is
+ *    necessary to install and execute a modified version of the
+ *    Combined Work produced by recombining or relinking the
+ *    Application with a modified version of the Linked Version. (If
+ *    you use option 4d0, the Installation Information must accompany
+ *    the Minimal Corresponding Source and Corresponding Application
+ *    Code. If you use option 4d1, you must provide the Installation
+ *    Information in the manner specified by section 6 of the GNU GPL
+ *    for conveying Corresponding Source.)
+ *
+ *   5. Combined Libraries.
+ *
+ *   You may place library facilities that are a work based on the
+ * Library side by side in a single library together with other library
+ * facilities that are not Applications and are not covered by this
+ * License, and convey such a combined library under terms of your
+ * choice, if you do both of the following:
+ *
+ *    a) Accompany the combined library with a copy of the same work based
+ *    on the Library, uncombined with any other library facilities,
+ *    conveyed under the terms of this License.
+ *
+ *    b) Give prominent notice with the combined library that part of it
+ *    is a work based on the Library, and explaining where to find the
+ *    accompanying uncombined form of the same work.
+ *
+ *   6. Revised Versions of the GNU Lesser General Public License.
+ *
+ *   The Free Software Foundation may publish revised and/or new versions
+ * of the GNU Lesser General Public License from time to time. Such new
+ * versions will be similar in spirit to the present version, but may
+ * differ in detail to address new problems or concerns.
+ *
+ *   Each version is given a distinguishing version number. If the
+ * Library as you received it specifies that a certain numbered version
+ * of the GNU Lesser General Public License "or any later version"
+ * applies to it, you have the option of following the terms and
+ * conditions either of that published version or of any later version
+ * published by the Free Software Foundation. If the Library as you
+ * received it does not specify a version number of the GNU Lesser
+ * General Public License, you may choose any version of the GNU Lesser
+ * General Public License ever published by the Free Software Foundation.
+ *
+ *   If the Library as you received it specifies that a proxy can decide
+ * whether future versions of the GNU Lesser General Public License shall
+ * apply, that proxy's public statement of acceptance of any version is
+ * permanent authorization for you to choose that version for the
+ * Library.
+ */
+
+package cc.hyperium.gui;
+
+import cc.hyperium.Hyperium;
+import cc.hyperium.config.Settings;
+import cc.hyperium.gui.hyperium.HyperiumMainGui;
+import cc.hyperium.handlers.handlers.SettingsMigrator;
+import cc.hyperium.mixinsimp.renderer.gui.IMixinGuiMultiplayer;
+import net.minecraft.client.Minecraft;
+import net.minecraft.client.gui.FontRenderer;
+import net.minecraft.client.gui.GuiButton;
+import net.minecraft.client.gui.GuiLanguage;
+import net.minecraft.client.gui.GuiMainMenu;
+import net.minecraft.client.gui.GuiMultiplayer;
+import net.minecraft.client.gui.GuiOptions;
+import net.minecraft.client.gui.GuiScreen;
+import net.minecraft.client.gui.GuiSelectWorld;
+import net.minecraft.client.gui.GuiYesNo;
+import net.minecraft.client.gui.GuiYesNoCallback;
+import net.minecraft.client.multiplayer.ServerData;
+import net.minecraft.client.renderer.GlStateManager;
+import net.minecraft.client.renderer.texture.DynamicTexture;
+import net.minecraft.client.renderer.texture.TextureManager;
+import net.minecraft.client.resources.I18n;
+import net.minecraft.client.settings.GameSettings;
+import net.minecraft.realms.RealmsBridge;
+import net.minecraft.util.ResourceLocation;
+import net.minecraft.world.demo.DemoWorldServer;
+import net.minecraft.world.storage.ISaveFormat;
+import net.minecraft.world.storage.WorldInfo;
+import org.lwjgl.input.Keyboard;
+
+import javax.imageio.ImageIO;
+import java.awt.Color;
+import java.io.IOException;
+import java.net.URL;
+import java.util.HashMap;
+
+public class GuiHyperiumScreenMainMenu extends GuiHyperiumScreen implements GuiYesNoCallback {
+
+
+    public static boolean FIRST_START = true;
+    private final ResourceLocation exit = new ResourceLocation("textures/material/exit.png");
+    private final ResourceLocation people_outline = new ResourceLocation("textures/material/people-outline.png");
+    private final ResourceLocation person_outline = new ResourceLocation("textures/material/person-outline.png");
+    private final ResourceLocation settings = new ResourceLocation("textures/material/settings.png");
+    private final ResourceLocation hIcon = new ResourceLocation("textures/h_icon.png");
+    private GuiScreen field_183503_M;
+    private boolean field_183502_L;
+    private int field_92021_u;
+    private int field_92022_t;
+    private int field_92024_r;
+    private int field_92023_s;
+    private FontRenderer fontRendererObj = Minecraft.getMinecraft().fontRendererObj;
+    private boolean clickedCheckBox = false;
+    private HashMap<String, DynamicTexture> cachedImages = new HashMap<>();
+
+    public GuiHyperiumScreenMainMenu() {
+        if (Minecraft.getMinecraft().isFullScreen() && Settings.WINDOWED_FULLSCREEN && FIRST_START) {
+            GuiHyperiumScreenMainMenu.FIRST_START = false;
+            Minecraft.getMinecraft().toggleFullscreen();
+            Minecraft.getMinecraft().toggleFullscreen();
+
+        }
+        if (Hyperium.INSTANCE.isFirstLaunch()) {
+            new SettingsMigrator().migrate();
+        }
+    }
+
+    /**
+     * Override initGui
+     *
+     * @author Cubxity
+     */
+
+    public void initGui() {
+
+        this.viewportTexture = new DynamicTexture(256, 256);
+        int j = this.height / 4 + 48;
+
+        this.addSingleplayerMultiplayerButtons(j - 10, 24);
+        this.buttonList.add(new GuiButton(100, this.width / 2 - 100, this.height - 45, I18n.format("button.menu.cosmeticshop")));
+
+        switch (getStyle()) {
+            case DEFAULT:
+                addDefaultStyleOptionsButton(j);
+                break;
+            case HYPERIUM:
+                addHyperiumStyleOptionsButton(j);
+                break;
+        }
+
+        this.mc.setConnectedToRealms(false);
+
+        if (Minecraft.getMinecraft().gameSettings.getOptionOrdinalValue(GameSettings.Options.REALMS_NOTIFICATIONS) && !this.field_183502_L) {
+            RealmsBridge realmsbridge = new RealmsBridge();
+            this.field_183503_M = realmsbridge.getNotificationScreen(this);
+            this.field_183502_L = true;
+        }
+
+        if (Minecraft.getMinecraft().gameSettings.getOptionOrdinalValue(GameSettings.Options.REALMS_NOTIFICATIONS) && this.field_183503_M != null) {
+            this.field_183503_M.func_183500_a(this.width, this.height);
+            this.field_183503_M.initGui();
+        }
+    }
+
+    /**
+     * Override buttons
+     *
+     * @author Cubxity
+     */
+
+    public void addSingleplayerMultiplayerButtons(int p_73969_1_, int p_73969_2_) {
+        switch (getStyle()) {
+            case DEFAULT:
+                addDefaultStyleSingleplayerMultiplayerButtons(p_73969_1_, p_73969_2_);
+                break;
+            case HYPERIUM:
+                addHyperiumStyleSingleplayerMultiplayerButtons(p_73969_1_, p_73969_2_);
+                break;
+        }
+    }
+
+    /**
+     * Override drawScreen method
+     *
+     * @author Cubxity
+     */
+    @Override
+    public void drawScreen(int mouseX, int mouseY, float partialTicks) {
+
+        super.drawScreen(mouseX, mouseY, partialTicks);
+
+    }
+
+    @Override
+    public void actionPerformed(GuiButton button) {
+
+        if (button.id == 0) {
+            this.mc.displayGuiScreen(new GuiOptions(this, this.mc.gameSettings));
+        }
+
+        if (button.id == 5) {
+            this.mc.displayGuiScreen(new GuiLanguage(this, this.mc.gameSettings, this.mc.getLanguageManager()));
+        }
+
+        if (button.id == 1) {
+            this.mc.displayGuiScreen(new GuiSelectWorld(this));
+        }
+
+        if (button.id == 2) {
+            this.mc.displayGuiScreen(new GuiMultiplayer(this));
+        }
+
+        if (button.id == 4) {
+            this.mc.shutdown();
+        }
+
+        if (button.id == 11) {
+            this.mc.launchIntegratedServer("Demo_World", "Demo_World", DemoWorldServer.demoWorldSettings);
+        }
+
+        if (button.id == 12) {
+            ISaveFormat isaveformat = this.mc.getSaveLoader();
+            WorldInfo worldinfo = isaveformat.getWorldInfo("Demo_World");
+
+            if (worldinfo != null) {
+                GuiYesNo guiyesno = GuiSelectWorld.func_152129_a(this, worldinfo.getWorldName(), 12);
+                this.mc.displayGuiScreen(guiyesno);
+            }
+        }
+
+        switch (getStyle()) {
+            case DEFAULT:
+                if (button.id == 15)
+                    HyperiumMainGui.INSTANCE.show();
+                if (button.id == 16) {
+                    GuiMultiplayer p_i1182_1_ = new GuiMultiplayer(new GuiMainMenu());
+                    p_i1182_1_.setWorldAndResolution(Minecraft.getMinecraft(), width, height);
+                    ((IMixinGuiMultiplayer) p_i1182_1_).makeDirectConnect();
+                    String hostName = Keyboard.isKeyDown(Keyboard.KEY_LSHIFT) ? "stuck.hypixel.net" : "mc.hypixel.net";
+                    ServerData data = new ServerData("hypixel", hostName, false);
+                    ((IMixinGuiMultiplayer) p_i1182_1_).setIp(data);
+                    p_i1182_1_.confirmClicked(true, 0);
+                }
+                if (button.id == 18) {
+
+                }
+                break;
+            case HYPERIUM:
+                if (button.id == 15)
+                    HyperiumMainGui.INSTANCE.show();
+                break;
+        }
+
+        if (button.id == 17)
+            mc.displayGuiScreen(new ChangeBackgroundGui(this));
+        if (button.id == 100) {
+            mc.displayGuiScreen(new ShopGui());
+        }
+
+    }
+
+    public void addHyperiumStyleSingleplayerMultiplayerButtons(int p_73969_1_, int p_73969_2_) {
+        this.buttonList.add(new GuiButton(1, this.width / 2 - getIntendedWidth(295), this.height / 2 - getIntendedHeight(55), getIntendedWidth(110), getIntendedHeight(110), ""));
+        this.buttonList.add(new GuiButton(2, this.width / 2 - getIntendedWidth(175), this.height / 2 - getIntendedHeight(55), getIntendedWidth(110), getIntendedHeight(110), ""));
+        this.buttonList.add(new GuiButton(15, this.width / 2 + getIntendedWidth(65), this.height / 2 - getIntendedHeight(55), getIntendedWidth(110), getIntendedHeight(110), ""));
+
+        // Background Gui Button
+        this.buttonList.add(new GuiButton(17, this.width / 2 - 110 / 2, this.height - 20, 110, getIntendedHeight(110), ""));
+    }
+
+    public void addDefaultStyleSingleplayerMultiplayerButtons(int p_73969_1_, int p_73969_2_) {
+        this.buttonList.add(new GuiButton(1, this.width / 2 - 100, p_73969_1_, I18n.format("menu.singleplayer")));
+        this.buttonList.add(new GuiButton(2, this.width / 2 - 100, p_73969_1_ + p_73969_2_, I18n.format("menu.multiplayer")));
+        //Change realms button ID to 16 to avoid conflicts
+        this.buttonList.add(this.hypixelButton = new GuiButton(16, this.width / 2 - 100, p_73969_1_ + p_73969_2_ * 2, 200, 20, "Join Hypixel"));
+
+        this.buttonList.add(new GuiButton(15, this.width / 2 - 100, p_73969_1_ + p_73969_2_ * 3, "Hyperium Settings"));
+
+        // Background Gui Button
+        this.buttonList.add(new GuiButton(17, this.width / 2 - 100, this.height - 20, "Change Menu Background"));
+    }
+
+    public void addHyperiumStyleOptionsButton(int j) {
+        this.buttonList.add(new GuiButton(0, width / 2 - getIntendedWidth(55), height / 2 - getIntendedHeight(55), getIntendedWidth(110), getIntendedHeight(110), ""));
+        this.buttonList.add(new GuiButton(4, width / 2 + getIntendedWidth(185), height / 2 - getIntendedHeight(55), getIntendedWidth(110), getIntendedHeight(110), ""));
+    }
+
+    public void addDefaultStyleOptionsButton(int j) {
+        this.buttonList.add(new GuiButton(0, this.width / 2 - 100, j + 72 + 12 + 24 - 5, 98, 20, I18n.format("menu.options")));
+        this.buttonList.add(new GuiButton(4, this.width / 2 + 2, j + 72 + 12 + 24 - 5, 98, 20, I18n.format("menu.quit")));
+    }
+
+    @Override
+    public void drawHyperiumStyleScreen(int mouseX, int mouseY, float partialTicks) {
+
+        GlStateManager.pushMatrix();
+
+        super.drawHyperiumStyleScreen(mouseX,mouseY,partialTicks);
+
+        sfr.drawCenteredString(I18n.format("button.menu.changebackground"), this.width / 2, this.height - 15, 0xFFFFFF);
+
+        // Draw icons on buttons
+        TextureManager tm = mc.getTextureManager();
+
+        tm.bindTexture(person_outline);
+        this.drawScaledCustomSizeModalRect(this.width / 2 - getIntendedWidth(285), this.height / 2 - getIntendedHeight(45), 0, 0, 192, 192, getIntendedWidth(90), getIntendedHeight(90), 192, 192);
+        tm.bindTexture(people_outline);
+        this.drawScaledCustomSizeModalRect(this.width / 2 - getIntendedWidth(165), this.height / 2 - getIntendedHeight(45), 0, 0, 192, 192, getIntendedWidth(90), getIntendedHeight(90), 192, 192);
+        tm.bindTexture(settings);
+        this.drawScaledCustomSizeModalRect(this.width / 2 - getIntendedWidth(45), this.height / 2 - getIntendedHeight(45), 0, 0, 192, 192, getIntendedWidth(90), getIntendedHeight(90), 192, 192);
+        tm.bindTexture(hIcon);
+        this.drawScaledCustomSizeModalRect(this.width / 2 + getIntendedWidth(85), this.height / 2 - getIntendedHeight(35), 0, 0, 104, 104, getIntendedWidth(70), getIntendedHeight(70), 104, 104);
+        tm.bindTexture(exit);
+        this.drawScaledCustomSizeModalRect(this.width / 2 + getIntendedWidth(195), this.height / 2 - getIntendedHeight(45), 0, 0, 192, 192, getIntendedWidth(90), getIntendedHeight(90), 192, 192);
+
+        GlStateManager.popMatrix();
+
+    }
+
+    private int color(int i, int i1, int i2, int i3) {
+        return new Color(i, i1, i2, i3).getRGB();
+    }
+
+    private DynamicTexture getCachedTexture(String t) {
+        final DynamicTexture[] texture = {this.cachedImages.get(t)};
+        if (texture[0] == null) {
+            Minecraft.getMinecraft().addScheduledTask(() -> {
+                try {
+                    texture[0] = new DynamicTexture(ImageIO
+                            .read(new URL("https://crafatar.com/avatars/" + t + "?size=30?default=MHF_Steve&overlay")));
+
+                } catch (Exception ignored) {
+                    try {
+                        texture[0] = new DynamicTexture(ImageIO
+                                .read(new URL("https://crafatar.com/avatars/c06f89064c8a49119c29ea1dbd1aab82")));
+                    } catch (IOException e) {
+                        e.printStackTrace();
+                    }
+                }
+                this.cachedImages.put(t, texture[0]);
+            });
+
+        }
+        return texture[0];
+    }
+
+    @Override
+    protected void keyTyped(char typedChar, int keyCode) throws IOException {
+        if (keyCode == Keyboard.KEY_B) {
+            Minecraft.getMinecraft().displayGuiScreen(new ChangeBackgroundGui(this));
+        }
+    }
+
+}
