--- src/main/java/cc/hyperium/gui/HyperiumMainMenu.java
@@ -1,859 +1,0 @@
-/*
- *                    GNU LESSER GENERAL PUBLIC LICENSE
- *                        Version 3, 29 June 2007
- *
- *  Copyright (C) 2007 Free Software Foundation, Inc. <http://fsf.org/>
- *  Everyone is permitted to copy and distribute verbatim copies
- *  of this license document, but changing it is not allowed.
- *
- *
- *   This version of the GNU Lesser General Public License incorporates
- * the terms and conditions of version 3 of the GNU General Public
- * License, supplemented by the additional permissions listed below.
- *
- *   0. Additional Definitions.
- *
- *   As used herein, "this License" refers to version 3 of the GNU Lesser
- * General Public License, and the "GNU GPL" refers to version 3 of the GNU
- * General Public License.
- *
- *   "The Library" refers to a covered work governed by this License,
- * other than an Application or a Combined Work as defined below.
- *
- *   An "Application" is any work that makes use of an interface provided
- * by the Library, but which is not otherwise based on the Library.
- * Defining a subclass of a class defined by the Library is deemed a mode
- * of using an interface provided by the Library.
- *
- *   A "Combined Work" is a work produced by combining or linking an
- * Application with the Library.  The particular version of the Library
- * with which the Combined Work was made is also called the "Linked
- * Version".
- *
- *   The "Minimal Corresponding Source" for a Combined Work means the
- * Corresponding Source for the Combined Work, excluding any source code
- * for portions of the Combined Work that, considered in isolation, are
- * based on the Application, and not on the Linked Version.
- *
- *   The "Corresponding Application Code" for a Combined Work means the
- * object code and/or source code for the Application, including any data
- * and utility programs needed for reproducing the Combined Work from the
- * Application, but excluding the System Libraries of the Combined Work.
- *
- *   1. Exception to Section 3 of the GNU GPL.
- *
- *   You may convey a covered work under sections 3 and 4 of this License
- * without being bound by section 3 of the GNU GPL.
- *
- *   2. Conveying Modified Versions.
- *
- *   If you modify a copy of the Library, and, in your modifications, a
- * facility refers to a function or data to be supplied by an Application
- * that uses the facility (other than as an argument passed when the
- * facility is invoked), then you may convey a copy of the modified
- * version:
- *
- *    a) under this License, provided that you make a good faith effort to
- *    ensure that, in the event an Application does not supply the
- *    function or data, the facility still operates, and performs
- *    whatever part of its purpose remains meaningful, or
- *
- *    b) under the GNU GPL, with none of the additional permissions of
- *    this License applicable to that copy.
- *
- *   3. Object Code Incorporating Material from Library Header Files.
- *
- *   The object code form of an Application may incorporate material from
- * a header file that is part of the Library.  You may convey such object
- * code under terms of your choice, provided that, if the incorporated
- * material is not limited to numerical parameters, data structure
- * layouts and accessors, or small macros, inline functions and templates
- * (ten or fewer lines in length), you do both of the following:
- *
- *    a) Give prominent notice with each copy of the object code that the
- *    Library is used in it and that the Library and its use are
- *    covered by this License.
- *
- *    b) Accompany the object code with a copy of the GNU GPL and this license
- *    document.
- *
- *   4. Combined Works.
- *
- *   You may convey a Combined Work under terms of your choice that,
- * taken together, effectively do not restrict modification of the
- * portions of the Library contained in the Combined Work and reverse
- * engineering for debugging such modifications, if you also do each of
- * the following:
- *
- *    a) Give prominent notice with each copy of the Combined Work that
- *    the Library is used in it and that the Library and its use are
- *    covered by this License.
- *
- *    b) Accompany the Combined Work with a copy of the GNU GPL and this license
- *    document.
- *
- *    c) For a Combined Work that displays copyright notices during
- *    execution, include the copyright notice for the Library among
- *    these notices, as well as a reference directing the user to the
- *    copies of the GNU GPL and this license document.
- *
- *    d) Do one of the following:
- *
- *        0) Convey the Minimal Corresponding Source under the terms of this
- *        License, and the Corresponding Application Code in a form
- *        suitable for, and under terms that permit, the user to
- *        recombine or relink the Application with a modified version of
- *        the Linked Version to produce a modified Combined Work, in the
- *        manner specified by section 6 of the GNU GPL for conveying
- *        Corresponding Source.
- *
- *        1) Use a suitable shared library mechanism for linking with the
- *        Library.  A suitable mechanism is one that (a) uses at run time
- *        a copy of the Library already present on the user's computer
- *        system, and (b) will operate properly with a modified version
- *        of the Library that is interface-compatible with the Linked
- *        Version.
- *
- *    e) Provide Installation Information, but only if you would otherwise
- *    be required to provide such information under section 6 of the
- *    GNU GPL, and only to the extent that such information is
- *    necessary to install and execute a modified version of the
- *    Combined Work produced by recombining or relinking the
- *    Application with a modified version of the Linked Version. (If
- *    you use option 4d0, the Installation Information must accompany
- *    the Minimal Corresponding Source and Corresponding Application
- *    Code. If you use option 4d1, you must provide the Installation
- *    Information in the manner specified by section 6 of the GNU GPL
- *    for conveying Corresponding Source.)
- *
- *   5. Combined Libraries.
- *
- *   You may place library facilities that are a work based on the
- * Library side by side in a single library together with other library
- * facilities that are not Applications and are not covered by this
- * License, and convey such a combined library under terms of your
- * choice, if you do both of the following:
- *
- *    a) Accompany the combined library with a copy of the same work based
- *    on the Library, uncombined with any other library facilities,
- *    conveyed under the terms of this License.
- *
- *    b) Give prominent notice with the combined library that part of it
- *    is a work based on the Library, and explaining where to find the
- *    accompanying uncombined form of the same work.
- *
- *   6. Revised Versions of the GNU Lesser General Public License.
- *
- *   The Free Software Foundation may publish revised and/or new versions
- * of the GNU Lesser General Public License from time to time. Such new
- * versions will be similar in spirit to the present version, but may
- * differ in detail to address new problems or concerns.
- *
- *   Each version is given a distinguishing version number. If the
- * Library as you received it specifies that a certain numbered version
- * of the GNU Lesser General Public License "or any later version"
- * applies to it, you have the option of following the terms and
- * conditions either of that published version or of any later version
- * published by the Free Software Foundation. If the Library as you
- * received it does not specify a version number of the GNU Lesser
- * General Public License, you may choose any version of the GNU Lesser
- * General Public License ever published by the Free Software Foundation.
- *
- *   If the Library as you received it specifies that a proxy can decide
- * whether future versions of the GNU Lesser General Public License shall
- * apply, that proxy's public statement of acceptance of any version is
- * permanent authorization for you to choose that version for the
- * Library.
- */
-
-package cc.hyperium.gui;
-
-import cc.hyperium.GuiStyle;
-import cc.hyperium.Hyperium;
-import cc.hyperium.Metadata;
-import cc.hyperium.config.Settings;
-import cc.hyperium.gui.hyperium.HyperiumMainGui;
-import cc.hyperium.gui.playerrenderer.GuiPlayerRenderer;
-import cc.hyperium.handlers.handlers.SettingsMigrator;
-import cc.hyperium.mixinsimp.renderer.gui.IMixinGuiMultiplayer;
-import cc.hyperium.purchases.PurchaseApi;
-import cc.hyperium.utils.HyperiumFontRenderer;
-import cc.hyperium.utils.JsonHolder;
-import com.mojang.realmsclient.gui.ChatFormatting;
-import net.minecraft.client.Minecraft;
-import net.minecraft.client.gui.FontRenderer;
-import net.minecraft.client.gui.GuiButton;
-import net.minecraft.client.gui.GuiLanguage;
-import net.minecraft.client.gui.GuiMainMenu;
-import net.minecraft.client.gui.GuiMultiplayer;
-import net.minecraft.client.gui.GuiOptions;
-import net.minecraft.client.gui.GuiScreen;
-import net.minecraft.client.gui.GuiSelectWorld;
-import net.minecraft.client.gui.GuiYesNo;
-import net.minecraft.client.gui.GuiYesNoCallback;
-import net.minecraft.client.gui.ScaledResolution;
-import net.minecraft.client.multiplayer.ServerData;
-import net.minecraft.client.renderer.GlStateManager;
-import net.minecraft.client.renderer.Tessellator;
-import net.minecraft.client.renderer.WorldRenderer;
-import net.minecraft.client.renderer.texture.DynamicTexture;
-import net.minecraft.client.renderer.texture.TextureManager;
-import net.minecraft.client.renderer.vertex.DefaultVertexFormats;
-import net.minecraft.client.resources.I18n;
-import net.minecraft.client.settings.GameSettings;
-import net.minecraft.realms.RealmsBridge;
-import net.minecraft.util.MathHelper;
-import net.minecraft.util.ResourceLocation;
-import net.minecraft.world.demo.DemoWorldServer;
-import net.minecraft.world.storage.ISaveFormat;
-import net.minecraft.world.storage.WorldInfo;
-import org.lwjgl.input.Keyboard;
-import org.lwjgl.opengl.GL11;
-import org.lwjgl.util.glu.Project;
-
-import javax.imageio.ImageIO;
-import java.awt.Color;
-import java.awt.Font;
-import java.awt.image.BufferedImage;
-import java.io.File;
-import java.io.FileInputStream;
-import java.io.IOException;
-import java.net.URL;
-import java.util.HashMap;
-
-public class HyperiumMainMenu extends GuiScreen implements GuiYesNoCallback {
-
-
-    public static boolean FIRST_START = true;
-    private static ResourceLocation background = new ResourceLocation("textures/material/backgrounds/1.png");
-    private static boolean customBackground = false;
-    private static File customImage = new File(Minecraft.getMinecraft().mcDataDir, "customImage.png");
-    private final ResourceLocation exit = new ResourceLocation("textures/material/exit.png");
-    private final ResourceLocation people_outline = new ResourceLocation("textures/material/people-outline.png");
-    private final ResourceLocation person_outline = new ResourceLocation("textures/material/person-outline.png");
-    private final ResourceLocation settings = new ResourceLocation("textures/material/settings.png");
-    private final ResourceLocation hIcon = new ResourceLocation("textures/h_icon.png");
-    private GuiScreen field_183503_M;
-    private boolean field_183502_L;
-    private int field_92021_u;
-    private int field_92022_t;
-    private int field_92024_r;
-    private int field_92023_s;
-    private DynamicTexture viewportTexture;
-    private FontRenderer fontRendererObj = Minecraft.getMinecraft().fontRendererObj;
-    private GuiButton hypixelButton;
-    private boolean clickedCheckBox = false;
-    private HyperiumFontRenderer fr = new HyperiumFontRenderer("Arial", Font.PLAIN, 20);
-    private HyperiumFontRenderer sfr = new HyperiumFontRenderer("Arial", Font.PLAIN, 12);
-    private HashMap<String, DynamicTexture> cachedImages = new HashMap<>();
-    private BufferedImage bgBr = null;
-    private ResourceLocation bgDynamicTexture = null;
-    private int panoramaTimer;
-    private static final ResourceLocation[] titlePanoramaPaths = new ResourceLocation[]{new ResourceLocation("textures/gui/title/background/panorama_0.png"), new ResourceLocation("textures/gui/title/background/panorama_1.png"), new ResourceLocation("textures/gui/title/background/panorama_2.png"), new ResourceLocation("textures/gui/title/background/panorama_3.png"), new ResourceLocation("textures/gui/title/background/panorama_4.png"), new ResourceLocation("textures/gui/title/background/panorama_5.png")};
-
-    private static float swing;
-
-    public HyperiumMainMenu() {
-        if (Minecraft.getMinecraft().isFullScreen() && Settings.WINDOWED_FULLSCREEN && FIRST_START) {
-            HyperiumMainMenu.FIRST_START = false;
-            Minecraft.getMinecraft().toggleFullscreen();
-            Minecraft.getMinecraft().toggleFullscreen();
-
-        }
-        if (Hyperium.INSTANCE.isFirstLaunch()) {
-            new SettingsMigrator().migrate();
-        }
-    }
-
-    public static ResourceLocation getBackground() {
-        return background;
-    }
-
-    public static void setBackground(ResourceLocation background) {
-        HyperiumMainMenu.background = background;
-    }
-
-    public static void setCustomBackground(boolean customBackground) {
-        HyperiumMainMenu.customBackground = customBackground;
-    }
-
-    /**
-     * Override initGui
-     *
-     * @author Cubxity
-     */
-
-    public void initGui() {
-        customBackground = Settings.BACKGROUND.equalsIgnoreCase("CUSTOM");
-        if (customImage.exists() && customBackground) {
-            try {
-                bgBr = ImageIO.read(new FileInputStream(customImage));
-            } catch (IOException e) {
-                e.printStackTrace();
-            }
-            if (bgBr != null)
-                bgDynamicTexture = mc.getRenderManager().renderEngine.getDynamicTextureLocation(customImage.getName(), new DynamicTexture(bgBr));
-            if (bgDynamicTexture == null)
-                return;
-        }
-
-        this.viewportTexture = new DynamicTexture(256, 256);
-        int j = this.height / 4 + 48;
-
-        this.addSingleplayerMultiplayerButtons(j - 10, 24);
-        this.buttonList.add(new GuiButton(100, this.width / 2 - 100, this.height - 45, I18n.format("button.menu.cosmeticshop")));
-
-        switch (getStyle()) {
-            case DEFAULT:
-                addDefaultStyleOptionsButton(j);
-                break;
-            case HYPERIUM:
-                addHyperiumStyleOptionsButton(j);
-                break;
-        }
-
-        this.mc.setConnectedToRealms(false);
-
-        if (Minecraft.getMinecraft().gameSettings.getOptionOrdinalValue(GameSettings.Options.REALMS_NOTIFICATIONS) && !this.field_183502_L) {
-            RealmsBridge realmsbridge = new RealmsBridge();
-            this.field_183503_M = realmsbridge.getNotificationScreen(this);
-            this.field_183502_L = true;
-        }
-
-        if (Minecraft.getMinecraft().gameSettings.getOptionOrdinalValue(GameSettings.Options.REALMS_NOTIFICATIONS) && this.field_183503_M != null) {
-            this.field_183503_M.func_183500_a(this.width, this.height);
-            this.field_183503_M.initGui();
-        }
-    }
-
-    /**
-     * Override buttons
-     *
-     * @author Cubxity
-     */
-
-    private void addSingleplayerMultiplayerButtons(int p_73969_1_, int p_73969_2_) {
-        switch (getStyle()) {
-            case DEFAULT:
-                addDefaultStyleSingleplayerMultiplayerButtons(p_73969_1_, p_73969_2_);
-                break;
-            case HYPERIUM:
-                addHyperiumStyleSingleplayerMultiplayerButtons(p_73969_1_, p_73969_2_);
-                break;
-        }
-    }
-
-    /**
-     * Override drawScreen method
-     *
-     * @author Cubxity
-     */
-
-    public void drawScreen(int mouseX, int mouseY, float partialTicks) {
-
-        switch (getStyle()) {
-            case DEFAULT:
-                drawDefaultStyleScreen(mouseX, mouseY, partialTicks);
-                break;
-            case HYPERIUM:
-                drawHyperiumStyleScreen(mouseX, mouseY, partialTicks);
-                break;
-        }
-
-        super.drawScreen(mouseX, mouseY, partialTicks);
-
-    }
-
-    @Override
-    public void actionPerformed(GuiButton button) {
-
-        if (button.id == 0) {
-            this.mc.displayGuiScreen(new GuiOptions(this, this.mc.gameSettings));
-        }
-
-        if (button.id == 5) {
-            this.mc.displayGuiScreen(new GuiLanguage(this, this.mc.gameSettings, this.mc.getLanguageManager()));
-        }
-
-        if (button.id == 1) {
-            this.mc.displayGuiScreen(new GuiSelectWorld(this));
-        }
-
-        if (button.id == 2) {
-            this.mc.displayGuiScreen(new GuiMultiplayer(this));
-        }
-
-        if (button.id == 4) {
-            this.mc.shutdown();
-        }
-
-        if (button.id == 11) {
-            this.mc.launchIntegratedServer("Demo_World", "Demo_World", DemoWorldServer.demoWorldSettings);
-        }
-
-        if (button.id == 12) {
-            ISaveFormat isaveformat = this.mc.getSaveLoader();
-            WorldInfo worldinfo = isaveformat.getWorldInfo("Demo_World");
-
-            if (worldinfo != null) {
-                GuiYesNo guiyesno = GuiSelectWorld.func_152129_a(this, worldinfo.getWorldName(), 12);
-                this.mc.displayGuiScreen(guiyesno);
-            }
-        }
-
-        switch (getStyle()) {
-            case DEFAULT:
-                if (button.id == 15)
-                    HyperiumMainGui.INSTANCE.show();
-                if (button.id == 16) {
-                    GuiMultiplayer p_i1182_1_ = new GuiMultiplayer(new GuiMainMenu());
-                    p_i1182_1_.setWorldAndResolution(Minecraft.getMinecraft(), width, height);
-                    ((IMixinGuiMultiplayer) p_i1182_1_).makeDirectConnect();
-                    String hostName = Keyboard.isKeyDown(Keyboard.KEY_LSHIFT) ? "stuck.hypixel.net" : "mc.hypixel.net";
-                    ServerData data = new ServerData("hypixel", hostName, false);
-                    ((IMixinGuiMultiplayer) p_i1182_1_).setIp(data);
-                    p_i1182_1_.confirmClicked(true, 0);
-                }
-                if (button.id == 18) {
-
-                }
-                break;
-            case HYPERIUM:
-                if (button.id == 15)
-                    HyperiumMainGui.INSTANCE.show();
-                break;
-        }
-
-        if (button.id == 17)
-            mc.displayGuiScreen(new ChangeBackgroundGui(this));
-        if (button.id == 100) {
-            mc.displayGuiScreen(new ShopGui());
-        }
-    }
-
-    private void addHyperiumStyleSingleplayerMultiplayerButtons(int p_73969_1_, int p_73969_2_) {
-        this.buttonList.add(new GuiButton(1, this.width / 2 - getIntendedWidth(295), this.height / 2 - getIntendedHeight(55), getIntendedWidth(110), getIntendedHeight(110), ""));
-        this.buttonList.add(new GuiButton(2, this.width / 2 - getIntendedWidth(175), this.height / 2 - getIntendedHeight(55), getIntendedWidth(110), getIntendedHeight(110), ""));
-        this.buttonList.add(new GuiButton(15, this.width / 2 + getIntendedWidth(65), this.height / 2 - getIntendedHeight(55), getIntendedWidth(110), getIntendedHeight(110), ""));
-
-        // Background Gui Button
-        this.buttonList.add(new GuiButton(17, this.width / 2 - 110 / 2, this.height - 20, 110, getIntendedHeight(110), ""));
-    }
-
-    private void addDefaultStyleSingleplayerMultiplayerButtons(int p_73969_1_, int p_73969_2_) {
-        this.buttonList.add(new GuiButton(1, this.width / 2 - 100, p_73969_1_, I18n.format("menu.singleplayer")));
-        this.buttonList.add(new GuiButton(2, this.width / 2 - 100, p_73969_1_ + p_73969_2_, I18n.format("menu.multiplayer")));
-        //Change realms button ID to 16 to avoid conflicts
-        this.buttonList.add(this.hypixelButton = new GuiButton(16, this.width / 2 - 100, p_73969_1_ + p_73969_2_ * 2, 200, 20, "Join Hypixel"));
-
-        this.buttonList.add(new GuiButton(15, this.width / 2 - 100, p_73969_1_ + p_73969_2_ * 3, "Hyperium Settings"));
-
-        // Background Gui Button
-        this.buttonList.add(new GuiButton(17, this.width / 2 - 100, this.height - 20, "Change Menu Background"));
-    }
-
-    private void addHyperiumStyleOptionsButton(int j) {
-        this.buttonList.add(new GuiButton(0, width / 2 - getIntendedWidth(55), height / 2 - getIntendedHeight(55), getIntendedWidth(110), getIntendedHeight(110), ""));
-        this.buttonList.add(new GuiButton(4, width / 2 + getIntendedWidth(185), height / 2 - getIntendedHeight(55), getIntendedWidth(110), getIntendedHeight(110), ""));
-    }
-
-    private void addDefaultStyleOptionsButton(int j) {
-        this.buttonList.add(new GuiButton(0, this.width / 2 - 100, j + 72 + 12 + 24 - 5, 98, 20, I18n.format("menu.options")));
-        this.buttonList.add(new GuiButton(4, this.width / 2 + 2, j + 72 + 12 + 24 - 5, 98, 20, I18n.format("menu.quit")));
-    }
-
-
-    private void drawHyperiumStyleScreen(int mouseX, int mouseY, float partialTicks) {
-
-        swing++;
-
-        if(Settings.BACKGROUND.equals("DEFAULT")) {
-            GlStateManager.disableAlpha();
-            this.renderSkybox(mouseX, mouseY, partialTicks);
-            GlStateManager.enableAlpha();
-        } else {
-            // Background
-            GlStateManager.disableAlpha();
-            ScaledResolution sr = new ScaledResolution(mc);
-            this.renderHyperiumBackground(sr);
-            GlStateManager.enableAlpha();
-
-        }
-
-        /* Render shadowed bar at top of screen */
-        this.drawGradientRect(0, 0, this.width, this.height, -2130706433, 16777215);
-        this.drawGradientRect(0, 0, this.width, this.height, 0, Integer.MIN_VALUE);
-
-        this.drawRect(0, 4, width,55, 0x66000000);
-        this.drawRect(0, 5, width,54, 0x66000000);
-
-        /* Render Client Logo */
-        GlStateManager.color(1,1,1,1);
-        ResourceLocation logo = new ResourceLocation("textures/hyperium-logo.png");
-        Minecraft.getMinecraft().getTextureManager().bindTexture(logo);
-        drawScaledCustomSizeModalRect(10, 5, 0, 0, 2160, 500, 200, 47, 2160, 500);
-
-        /* Render profile container */
-        this.drawRect(width - 155, 10, width - 10, 49, 0x33000000);
-        this.drawRect(width - 156, 9, width - 9, 50, 0x33000000);
-
-        /* Render client version container */
-        /*
-        drawRect(width - 153, 39, width - 60, 48, 0x33000000);
-        drawRect(width - 154, 40, width - 59, 49, 0x33000000);
-        */
-
-        // Looks weird with the small green strip
-        // drawRect(width - 160, 10, width - 158, 40, new Color(149, 201, 144, 255).getRGB());
-
-        /* Fetch player credit count */
-        JsonHolder response = PurchaseApi.getInstance().getSelf().getResponse();
-        int credits = response.optInt("total_credits");
-
-        /* Render player credits count and username */
-        fr.drawString(Minecraft.getMinecraft().getSession().getUsername(), width - 153, 13, 0xFFFFFF);
-        fr.drawString(I18n.format("menu.profile.credits", credits), width - 153, 25, 0xFFFF00);
-
-
-        float val = (float) (Math.sin(swing / 40) * 30);
-
-        ScissorState.scissor(width - 153, 0,145,49, true);
-        GuiPlayerRenderer.renderPlayerWithRotation(width - 118,-4,val);
-        ScissorState.endScissor();
-
-        /* Render Hyperium version number */
-        fr.drawStringScaled("Hyperium v" + (Hyperium.INSTANCE.isLatestVersion ? ChatFormatting.GREEN : ChatFormatting.RED) + Metadata.getVersion(),width - 152, 39, 0xFFFFFF,.75);
-
-        /* Display profile player skin head */
-        /*
-        GlStateManager.pushMatrix();
-        GlStateManager.enableBlend();
-        GL11.glTexParameteri(GL11.GL_TEXTURE_2D, GL11.GL_TEXTURE_MAG_FILTER, GL11.GL_NEAREST);
-        GlStateManager.bindTexture(getCachedTexture(Minecraft.getMinecraft().getSession().getPlayerID()).getGlTextureId());
-        drawScaledCustomSizeModalRect(width - 153, 12, 0, 0, 30, 30, 23, 25, 30, 30);
-        */
-
-        /* Display copyright disclaimers at bottom of screen */
-        sfr.drawString(I18n.format("menu.left").toUpperCase(), 1, height - 7, 0x55FFFFFF);
-        String s = I18n.format("menu.right").toUpperCase();
-        sfr.drawString(s, width - sfr.getWidth(s) - 1, height - 7, 0x55FFFFFF);
-
-        sfr.drawCenteredString(I18n.format("button.menu.changebackground"), this.width / 2, this.height - 15, 0xFFFFFF);
-
-        // Draw icons on buttons
-        TextureManager tm = mc.getTextureManager();
-
-        tm.bindTexture(person_outline);
-        this.drawScaledCustomSizeModalRect(this.width / 2 - getIntendedWidth(285), this.height / 2 - getIntendedHeight(45), 0, 0, 192, 192, getIntendedWidth(90), getIntendedHeight(90), 192, 192);
-        tm.bindTexture(people_outline);
-        this.drawScaledCustomSizeModalRect(this.width / 2 - getIntendedWidth(165), this.height / 2 - getIntendedHeight(45), 0, 0, 192, 192, getIntendedWidth(90), getIntendedHeight(90), 192, 192);
-        tm.bindTexture(settings);
-        this.drawScaledCustomSizeModalRect(this.width / 2 - getIntendedWidth(45), this.height / 2 - getIntendedHeight(45), 0, 0, 192, 192, getIntendedWidth(90), getIntendedHeight(90), 192, 192);
-        tm.bindTexture(hIcon);
-        this.drawScaledCustomSizeModalRect(this.width / 2 + getIntendedWidth(85), this.height / 2 - getIntendedHeight(35), 0, 0, 104, 104, getIntendedWidth(70), getIntendedHeight(70), 104, 104);
-        tm.bindTexture(exit);
-        this.drawScaledCustomSizeModalRect(this.width / 2 + getIntendedWidth(195), this.height / 2 - getIntendedHeight(45), 0, 0, 192, 192, getIntendedWidth(90), getIntendedHeight(90), 192, 192);
-        GlStateManager.disableBlend();
-        GlStateManager.popMatrix();
-    }
-
-    private int getIntendedWidth(int value) {
-        float intendedWidth = 1920F;
-        return (int) ((Minecraft.getMinecraft().displayWidth / intendedWidth) * value);
-    }
-
-    private int getIntendedHeight(int value) {
-        float intendedHeight = 1080F;
-        return (int) ((Minecraft.getMinecraft().displayHeight / intendedHeight) * value);
-    }
-
-    private void renderHyperiumBackground(ScaledResolution p_180476_1_) {
-        GlStateManager.disableDepth();
-        GlStateManager.depthMask(false);
-        GlStateManager.tryBlendFuncSeparate(770, 771, 1, 0);
-        GlStateManager.color(1.0F, 1.0F, 1.0F, 1.0F);
-        GlStateManager.disableAlpha();
-        if (customImage.exists() && bgDynamicTexture != null && customBackground) {
-            Minecraft.getMinecraft().getTextureManager().bindTexture(bgDynamicTexture);
-            Tessellator tessellator = Tessellator.getInstance();
-            WorldRenderer worldrenderer = tessellator.getWorldRenderer();
-            worldrenderer.begin(7, DefaultVertexFormats.POSITION_TEX);
-            worldrenderer.pos(0.0D, (double) p_180476_1_.getScaledHeight(), -90.0D).tex(0.0D, 1.0D).endVertex();
-            worldrenderer.pos((double) p_180476_1_.getScaledWidth(), (double) p_180476_1_.getScaledHeight(), -90.0D).tex(1.0D, 1.0D).endVertex();
-            worldrenderer.pos((double) p_180476_1_.getScaledWidth(), 0.0D, -90.0D).tex(1.0D, 0.0D).endVertex();
-            worldrenderer.pos(0.0D, 0.0D, -90.0D).tex(0.0D, 0.0D).endVertex();
-            tessellator.draw();
-
-            GlStateManager.depthMask(true);
-            GlStateManager.enableDepth();
-            GlStateManager.enableAlpha();
-            GlStateManager.color(1.0F, 1.0F, 1.0F, 1.0F);
-            return;
-        }
-        Minecraft.getMinecraft().getTextureManager().bindTexture(background);
-        Tessellator tessellator = Tessellator.getInstance();
-        WorldRenderer worldrenderer = tessellator.getWorldRenderer();
-        worldrenderer.begin(7, DefaultVertexFormats.POSITION_TEX);
-        worldrenderer.pos(0.0D, (double) p_180476_1_.getScaledHeight(), -90.0D).tex(0.0D, 1.0D).endVertex();
-        worldrenderer.pos((double) p_180476_1_.getScaledWidth(), (double) p_180476_1_.getScaledHeight(), -90.0D).tex(1.0D, 1.0D).endVertex();
-        worldrenderer.pos((double) p_180476_1_.getScaledWidth(), 0.0D, -90.0D).tex(1.0D, 0.0D).endVertex();
-        worldrenderer.pos(0.0D, 0.0D, -90.0D).tex(0.0D, 0.0D).endVertex();
-        tessellator.draw();
-        GlStateManager.depthMask(true);
-        GlStateManager.enableDepth();
-        GlStateManager.enableAlpha();
-        GlStateManager.color(1.0F, 1.0F, 1.0F, 1.0F);
-    }
-
-
-    private int color(int i, int i1, int i2, int i3) {
-        return new Color(i, i1, i2, i3).getRGB();
-    }
-
-    private void drawDefaultStyleScreen(int mouseX, int mouseY, float partialTicks) {
-        if(Settings.BACKGROUND.equals("DEFAULT")) {
-            GlStateManager.disableAlpha();
-            this.renderSkybox(mouseX, mouseY, partialTicks);
-            GlStateManager.enableAlpha();
-        } else {
-            // Background
-            GlStateManager.disableAlpha();
-            this.renderHyperiumBackground(new ScaledResolution(mc));
-
-        }
-        ParticleOverlay.getOverlay().render(mouseX, mouseY, 0, 0, 0, 0);
-
-        GlStateManager.enableAlpha();
-        this.drawGradientRect(0, 0, this.width, this.height, -2130706433, 16777215);
-        this.drawGradientRect(0, 0, this.width, this.height, 0, Integer.MIN_VALUE);
-        GlStateManager.pushMatrix();
-        GlStateManager.scale(4F, 4F, 1F);
-        this.drawCenteredString(fontRendererObj, Metadata.getModid(), width / 8, 40 / 4, 0xFFFFFF);
-        GlStateManager.popMatrix();
-        String s = String.format("%s %s", Metadata.getModid(), Metadata.getVersion());
-        this.drawString(this.fontRendererObj, s, 2, this.height - 10, -1);
-        String s1 = "Not affiliated with Mojang AB.";
-        this.drawString(this.fontRendererObj, s1, this.width - this.fontRendererObj.getStringWidth(s1) - 2, this.height - 10, -1);
-        String s3 = "Made by Sk1er, Kevin,";
-        this.drawString(this.fontRendererObj, s3, this.width - this.fontRendererObj.getStringWidth(s3) - 2, this.height - 30, -1);
-
-        String s4 = "Cubxity, CoalOres and boomboompower";
-        this.drawString(this.fontRendererObj, s4, this.width - this.fontRendererObj.getStringWidth(s4) - 2, this.height - 20, -1);
-        GuiButton hypixelButton = this.hypixelButton;
-        if (hypixelButton != null)
-            hypixelButton.displayString = Keyboard.isKeyDown(Keyboard.KEY_LSHIFT) ? "Fix Hypixel Session" : "Join Hypixel";
-    }
-
-    /**
-     * Rotate and blurs the skybox view in the main menu
-     */
-    private void rotateAndBlurSkybox(float p_73968_1_) {
-        ResourceLocation backgroundTexture = this.mc.getTextureManager().getDynamicTextureLocation("background", this.viewportTexture);
-        this.mc.getTextureManager().bindTexture(backgroundTexture);
-        GL11.glTexParameteri(GL11.GL_TEXTURE_2D, GL11.GL_TEXTURE_MIN_FILTER, GL11.GL_LINEAR);
-        GL11.glTexParameteri(GL11.GL_TEXTURE_2D, GL11.GL_TEXTURE_MAG_FILTER, GL11.GL_LINEAR);
-        GL11.glCopyTexSubImage2D(GL11.GL_TEXTURE_2D, 0, 0, 0, 0, 0, 256, 256);
-        GlStateManager.enableBlend();
-        GlStateManager.tryBlendFuncSeparate(770, 771, 1, 0);
-        GlStateManager.colorMask(true, true, true, false);
-        Tessellator tessellator = Tessellator.getInstance();
-        WorldRenderer worldrenderer = tessellator.getWorldRenderer();
-        worldrenderer.begin(7, DefaultVertexFormats.POSITION_TEX_COLOR);
-        GlStateManager.disableAlpha();
-        int i = 3;
-
-        for (int j = 0; j < i; ++j)
-        {
-            float f = 1.0F / (float)(j + 1);
-            int k = this.width;
-            int l = this.height;
-            float f1 = (float)(j - i / 2) / 256.0F;
-            worldrenderer.pos((double)k, (double)l, (double)this.zLevel).tex((double)(0.0F + f1), 1.0D).color(1.0F, 1.0F, 1.0F, f).endVertex();
-            worldrenderer.pos((double)k, 0.0D, (double)this.zLevel).tex((double)(1.0F + f1), 1.0D).color(1.0F, 1.0F, 1.0F, f).endVertex();
-            worldrenderer.pos(0.0D, 0.0D, (double)this.zLevel).tex((double)(1.0F + f1), 0.0D).color(1.0F, 1.0F, 1.0F, f).endVertex();
-            worldrenderer.pos(0.0D, (double)l, (double)this.zLevel).tex((double)(0.0F + f1), 0.0D).color(1.0F, 1.0F, 1.0F, f).endVertex();
-        }
-
-        tessellator.draw();
-        GlStateManager.enableAlpha();
-        GlStateManager.colorMask(true, true, true, true);
-    }
-
-    /**
-     * Renders the skybox in the main menu
-     */
-    private void renderSkybox(int p_73971_1_, int p_73971_2_, float p_73971_3_)
-    {
-        this.mc.getFramebuffer().unbindFramebuffer();
-        GlStateManager.viewport(0, 0, 256, 256);
-        this.drawPanorama(p_73971_1_, p_73971_2_, p_73971_3_);
-        this.rotateAndBlurSkybox(p_73971_3_);
-        this.rotateAndBlurSkybox(p_73971_3_);
-        this.rotateAndBlurSkybox(p_73971_3_);
-        this.rotateAndBlurSkybox(p_73971_3_);
-        this.rotateAndBlurSkybox(p_73971_3_);
-        this.rotateAndBlurSkybox(p_73971_3_);
-        this.rotateAndBlurSkybox(p_73971_3_);
-        this.mc.getFramebuffer().bindFramebuffer(true);
-        GlStateManager.viewport(0, 0, this.mc.displayWidth, this.mc.displayHeight);
-        float f = this.width > this.height ? 120.0F / (float)this.width : 120.0F / (float)this.height;
-        float f1 = (float)this.height * f / 256.0F;
-        float f2 = (float)this.width * f / 256.0F;
-        int i = this.width;
-        int j = this.height;
-        Tessellator tessellator = Tessellator.getInstance();
-        WorldRenderer worldrenderer = tessellator.getWorldRenderer();
-        worldrenderer.begin(7, DefaultVertexFormats.POSITION_TEX_COLOR);
-        worldrenderer.pos(0.0D, (double)j, (double)this.zLevel).tex((double)(0.5F - f1), (double)(0.5F + f2)).color(1.0F, 1.0F, 1.0F, 1.0F).endVertex();
-        worldrenderer.pos((double)i, (double)j, (double)this.zLevel).tex((double)(0.5F - f1), (double)(0.5F - f2)).color(1.0F, 1.0F, 1.0F, 1.0F).endVertex();
-        worldrenderer.pos((double)i, 0.0D, (double)this.zLevel).tex((double)(0.5F + f1), (double)(0.5F - f2)).color(1.0F, 1.0F, 1.0F, 1.0F).endVertex();
-        worldrenderer.pos(0.0D, 0.0D, (double)this.zLevel).tex((double)(0.5F + f1), (double)(0.5F + f2)).color(1.0F, 1.0F, 1.0F, 1.0F).endVertex();
-        tessellator.draw();
-    }
-
-
-    private void drawPanorama(int p_73970_1_, int p_73970_2_, float p_73970_3_)
-    {
-        Tessellator tessellator = Tessellator.getInstance();
-        WorldRenderer worldrenderer = tessellator.getWorldRenderer();
-        GlStateManager.matrixMode(5889);
-        GlStateManager.pushMatrix();
-        GlStateManager.loadIdentity();
-        Project.gluPerspective(120.0F, 1.0F, 0.05F, 10.0F);
-        GlStateManager.matrixMode(5888);
-        GlStateManager.pushMatrix();
-        GlStateManager.loadIdentity();
-        GlStateManager.color(1.0F, 1.0F, 1.0F, 1.0F);
-        GlStateManager.rotate(180.0F, 1.0F, 0.0F, 0.0F);
-        GlStateManager.rotate(90.0F, 0.0F, 0.0F, 1.0F);
-        GlStateManager.enableBlend();
-        GlStateManager.disableAlpha();
-        GlStateManager.disableCull();
-        GlStateManager.depthMask(false);
-        GlStateManager.tryBlendFuncSeparate(770, 771, 1, 0);
-        int i = 8;
-
-        for (int j = 0; j < i * i; ++j)
-        {
-            GlStateManager.pushMatrix();
-            float f = ((float)(j % i) / (float)i - 0.5F) / 64.0F;
-            float f1 = ((float)(j / i) / (float)i - 0.5F) / 64.0F;
-            float f2 = 0.0F;
-            GlStateManager.translate(f, f1, f2);
-            GlStateManager.rotate(MathHelper.sin(((float)this.panoramaTimer + p_73970_3_) / 400.0F) * 25.0F + 20.0F, 1.0F, 0.0F, 0.0F);
-            GlStateManager.rotate(-((float)this.panoramaTimer + p_73970_3_) * 0.1F, 0.0F, 1.0F, 0.0F);
-
-            for (int k = 0; k < 6; ++k)
-            {
-                GlStateManager.pushMatrix();
-
-                if (k == 1)
-                {
-                    GlStateManager.rotate(90.0F, 0.0F, 1.0F, 0.0F);
-                }
-
-                if (k == 2)
-                {
-                    GlStateManager.rotate(180.0F, 0.0F, 1.0F, 0.0F);
-                }
-
-                if (k == 3)
-                {
-                    GlStateManager.rotate(-90.0F, 0.0F, 1.0F, 0.0F);
-                }
-
-                if (k == 4)
-                {
-                    GlStateManager.rotate(90.0F, 1.0F, 0.0F, 0.0F);
-                }
-
-                if (k == 5)
-                {
-                    GlStateManager.rotate(-90.0F, 1.0F, 0.0F, 0.0F);
-                }
-
-                this.mc.getTextureManager().bindTexture(titlePanoramaPaths[k]);
-                worldrenderer.begin(7, DefaultVertexFormats.POSITION_TEX_COLOR);
-                int l = 255 / (j + 1);
-                float f3 = 0.0F;
-                worldrenderer.pos(-1.0D, -1.0D, 1.0D).tex(0.0D, 0.0D).color(255, 255, 255, l).endVertex();
-                worldrenderer.pos(1.0D, -1.0D, 1.0D).tex(1.0D, 0.0D).color(255, 255, 255, l).endVertex();
-                worldrenderer.pos(1.0D, 1.0D, 1.0D).tex(1.0D, 1.0D).color(255, 255, 255, l).endVertex();
-                worldrenderer.pos(-1.0D, 1.0D, 1.0D).tex(0.0D, 1.0D).color(255, 255, 255, l).endVertex();
-                tessellator.draw();
-                GlStateManager.popMatrix();
-            }
-
-            GlStateManager.popMatrix();
-            GlStateManager.colorMask(true, true, true, false);
-        }
-
-        worldrenderer.setTranslation(0.0D, 0.0D, 0.0D);
-        GlStateManager.colorMask(true, true, true, true);
-        GlStateManager.matrixMode(5889);
-        GlStateManager.popMatrix();
-        GlStateManager.matrixMode(5888);
-        GlStateManager.popMatrix();
-        GlStateManager.depthMask(true);
-        GlStateManager.enableCull();
-        GlStateManager.enableDepth();
-    }
-
-    private GuiStyle getStyle() {
-        return GuiStyle.valueOf(Settings.MENU_STYLE);
-    }
-
-    private DynamicTexture getCachedTexture(String t) {
-        final DynamicTexture[] texture = {this.cachedImages.get(t)};
-        if (texture[0] == null) {
-            Minecraft.getMinecraft().addScheduledTask(() -> {
-                try {
-                    texture[0] = new DynamicTexture(ImageIO
-                            .read(new URL("https://crafatar.com/avatars/" + t + "?size=30?default=MHF_Steve&overlay")));
-
-                } catch (Exception ignored) {
-                    try {
-                        texture[0] = new DynamicTexture(ImageIO
-                                .read(new URL("https://crafatar.com/avatars/c06f89064c8a49119c29ea1dbd1aab82")));
-                    } catch (IOException e) {
-                        e.printStackTrace();
-                    }
-                }
-                this.cachedImages.put(t, texture[0]);
-            });
-
-        }
-        return texture[0];
-    }
-
-    @Override
-    protected void keyTyped(char typedChar, int keyCode) throws IOException {
-        if (keyCode == Keyboard.KEY_B) {
-            Minecraft.getMinecraft().displayGuiScreen(new ChangeBackgroundGui(this));
-        }
-    }
-
-    public static void drawScaledCustomSizeModalRect(int x, int y, float u, float v, int uWidth, int vHeight, int width, int height, float tileWidth, float tileHeight)
-    {
-        float f = 1.0F / tileWidth;
-        float f1 = 1.0F / tileHeight;
-
-        Tessellator tessellator = Tessellator.getInstance();
-        WorldRenderer worldrenderer = tessellator.getWorldRenderer();
-
-        GL11.glEnable(GL11.GL_BLEND);
-        GL11.glEnable(GL11.GL_POINT_SMOOTH);
-        GL11.glHint(GL11.GL_POINT_SMOOTH_HINT, GL11.GL_FASTEST);
-
-        worldrenderer.begin(7, DefaultVertexFormats.POSITION_TEX);
-        worldrenderer.pos((double)x, (double)(y + height), 0.0D).tex((double)(u * f), (double)((v + (float)vHeight) * f1)).endVertex();
-        worldrenderer.pos((double)(x + width), (double)(y + height), 0.0D).tex((double)((u + (float)uWidth) * f), (double)((v + (float)vHeight) * f1)).endVertex();
-        worldrenderer.pos((double)(x + width), (double)y, 0.0D).tex((double)((u + (float)uWidth) * f), (double)(v * f1)).endVertex();
-        worldrenderer.pos((double)x, (double)y, 0.0D).tex((double)(u * f), (double)(v * f1)).endVertex();
-        tessellator.draw();
-
-        GL11.glDisable(GL11.GL_BLEND);
-        GL11.glDisable(GL11.GL_POINT_SMOOTH);
-    }
-
-}
